+++ libs/base_alloc/src/work/sync/mod.rs
@@ -7,12 +7,8 @@
 // safety
 #![cfg_attr(base_safe_work, forbid(unsafe_code))]
 
-mod reexports;
+mod _reexport;
 
-crate::structural_mods! { // _mods
-    _mods {
-        pub use super::{
-            reexports::*,
-        };
-    }
+crate::structural_mods! { // _reexports
+    _reexports { pub use super::_reexport::*; }
 }