+++ libs/base_core/src/work/sync/atomic/_reexport.rs
@@ -4,25 +4,23 @@
 //! and the [`Atomic`] type from the [atomic](https://docs.rs/atomic) crate.
 //
 
-use crate::{_TAG_ATOMIC, _reexport};
-
-/* from `core` */
+use crate::{_TAG_ATOMIC, _TAG_CONCURRENCY, _reexport};
 
 // enums
-_reexport! { rust: core::sync::atomic,
-    tag: _TAG_ATOMIC!(),
+_reexport! { rust: core::sync::atomic, location: "work/sync/atomic",
+    tag: _TAG_CONCURRENCY!() _TAG_ATOMIC!(),
     doc: "Atomic memory ordering.",
     @Ordering as AtomicOrdering
 }
 
 // functions
-_reexport! { rust: core::sync::atomic,
-    tag: _TAG_ATOMIC!(),
+_reexport! { rust: core::sync::atomic, location: "work/sync/atomic",
+    tag: _TAG_CONCURRENCY!() _TAG_ATOMIC!(),
     doc: "An atomic fence.",
     @fence as atomic_fence
 }
-_reexport! { rust: core::sync::atomic,
-    tag: _TAG_ATOMIC!(),
+_reexport! { rust: core::sync::atomic, location: "work/sync/atomic",
+    tag: _TAG_CONCURRENCY!() _TAG_ATOMIC!(),
     doc: "A compiler memory fence.",
     @compiler_fence as atomic_compiler_fence
 }