+++ src/work/future/ext.rs
@@ -32,7 +32,10 @@ See [arbitrary_self_types](https://github.com/rust-lang/rust/issues/44874).
 "#;
 }
 
+#[doc = crate::_TAG_CONCURRENCY!()]
+#[doc = crate::_TAG_RUNTIME!()]
 /// Extension trait providing additional methods for [`Future`]s.
+#[doc = crate::_doc_location!("work/future")]
 #[rustfmt::skip]
 pub trait FutureExt: Future {
     #[doc = DOC_BLOCK_ON!()]