+++ src/work/sync/atomic/_reexport_dep.rs
@@ -6,12 +6,13 @@
 
 #[cfg(feature = "dep_portable_atomic")]
 use crate::_DOC_ATOMIC_CORE_PORTABLE;
-use crate::{_TAG_ATOMIC, _TAG_ATOMIC_CORE_PORTABLE, _reexport};
+use crate::{_TAG_ATOMIC, _TAG_ATOMIC_CORE_PORTABLE, _TAG_CONCURRENCY, _reexport};
 
 /* from the `atomic` crate */
 
 _reexport! { "dep_atomic", "atomic", atomic,
-    tag: _TAG_ATOMIC!(),
+    location: "work/sync/atomic",
+    tag: _TAG_CONCURRENCY!() _TAG_ATOMIC!(),
     doc: "A generic atomic wrapper type.",
     Atomic
 }
@@ -19,55 +20,63 @@ _reexport! { "dep_atomic", "atomic", atomic,
 /* from `portable-atomic` */
 
 _reexport! { "dep_portable_atomic", "portable-atomic", portable_atomic,
-    tag: _TAG_ATOMIC!(),
+    location: "work/sync/atomic", tag: _TAG_CONCURRENCY!() _TAG_ATOMIC!(),
     doc: "A thread-safe floating-point type.",
     AtomicF32, AtomicF64
 }
 _reexport! { "dep_portable_atomic", "portable-atomic", portable_atomic,
-    tag: _TAG_ATOMIC!(),
+    location: "work/sync/atomic", tag:  _TAG_CONCURRENCY!()_TAG_ATOMIC!(),
     doc: "A thread-safe signed integer type.",
     AtomicI128
 }
 _reexport! { "dep_portable_atomic", "portable-atomic", portable_atomic,
-    tag: _TAG_ATOMIC!(),
+    location: "work/sync/atomic", tag:  _TAG_CONCURRENCY!()_TAG_ATOMIC!(),
     doc: "A thread-safe unsigned integer type.",
     AtomicU128
 }
 
 /* from either `portable-atomic` or `core` */
 
+#[doc = _TAG_CONCURRENCY!()]
 #[doc = _TAG_ATOMIC!()]
 #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
-#[doc = "A thread-safe signed integer type.\n\n"]
+/// A thread-safe signed integer type.
+#[doc = crate::_doc_location!("work/sync/atomic")]
 #[doc = _DOC_ATOMIC_CORE_PORTABLE!()]
 #[cfg(feature = "dep_portable_atomic")]
 pub use crate::_dep::portable_atomic::{AtomicI8, AtomicI16, AtomicI32, AtomicI64, AtomicIsize};
 
+#[doc = _TAG_CONCURRENCY!()]
 #[doc = _TAG_ATOMIC!()]
 #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
-#[doc = "A thread-safe unsigned integer type.\n\n"]
+/// A thread-safe signed integer type.
+#[doc = crate::_doc_location!("work/sync/atomic")]
 #[doc = _DOC_ATOMIC_CORE_PORTABLE!()]
 #[cfg(feature = "dep_portable_atomic")]
 pub use crate::_dep::portable_atomic::{AtomicU8, AtomicU16, AtomicU32, AtomicU64, AtomicUsize};
 
+#[doc = _TAG_CONCURRENCY!()]
 #[doc = _TAG_ATOMIC!()]
 #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
 #[cfg(all(not(feature = "dep_portable_atomic"), target_has_atomic = "8"))]
 #[cfg_attr(nightly_doc, doc(cfg(target_has_atomic = "8")))]
 pub use core::sync::atomic::{AtomicI8, AtomicU8};
 //
+#[doc = _TAG_CONCURRENCY!()]
 #[doc = _TAG_ATOMIC!()]
 #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
 #[cfg(all(not(feature = "dep_portable_atomic"), target_has_atomic = "16"))]
 #[cfg_attr(nightly_doc, doc(cfg(target_has_atomic = "16")))]
 pub use core::sync::atomic::{AtomicI16, AtomicU16};
 //
+#[doc = _TAG_CONCURRENCY!()]
 #[doc = _TAG_ATOMIC!()]
 #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
 #[cfg(all(not(feature = "dep_portable_atomic"), target_has_atomic = "32"))]
 #[cfg_attr(nightly_doc, doc(cfg(target_has_atomic = "32")))]
 pub use core::sync::atomic::{AtomicI32, AtomicU32};
 //
+#[doc = _TAG_CONCURRENCY!()]
 #[doc = _TAG_ATOMIC!()]
 #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
 #[cfg(all(not(feature = "dep_portable_atomic"), target_has_atomic = "64"))]
@@ -75,37 +84,51 @@ pub use core::sync::atomic::{AtomicI32, AtomicU32};
 pub use core::sync::atomic::{AtomicI64, AtomicU64};
 //
 // WAIT: [integer_atomics](https://github.com/rust-lang/rust/issues/99069)
+// #[doc = _TAG_CONCURRENCY!()]
 // #[doc = _TAG_ATOMIC!()]
 // #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
 // #[cfg(all(not(feature = "dep_portable_atomic"), target_has_atomic = "128"))]
 // #[cfg_attr(nightly_doc, doc(cfg(target_has_atomic = "128")))]
 // pub use core::sync::atomic::{AtomicI128, AtomicU128};
 //
+#[doc = _TAG_CONCURRENCY!()]
 #[doc = _TAG_ATOMIC!()]
 #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
 #[cfg(all(not(feature = "dep_portable_atomic"), target_has_atomic = "ptr"))]
 #[cfg_attr(nightly_doc, doc(cfg(target_has_atomic = "ptr")))]
 pub use core::sync::atomic::{AtomicIsize, AtomicUsize};
 
+#[doc = _TAG_CONCURRENCY!()]
 #[doc = _TAG_ATOMIC!()]
 #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
+/// A raw pointer type which can be safely shared between threads.
+#[doc = crate::_doc_location!("work/sync/atomic")]
 #[cfg(feature = "dep_portable_atomic")]
 pub use crate::_dep::portable_atomic::AtomicPtr;
 //
+#[doc = _TAG_CONCURRENCY!()]
 #[doc = _TAG_ATOMIC!()]
 #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
+/// A raw pointer type which can be safely shared between threads.
+#[doc = crate::_doc_location!("work/sync/atomic")]
 #[cfg(all(not(feature = "dep_portable_atomic"), target_has_atomic = "ptr"))]
 #[cfg_attr(nightly_doc, doc(cfg(target_has_atomic = "ptr")))]
 pub use core::sync::atomic::AtomicPtr;
 
+#[doc = _TAG_CONCURRENCY!()]
 #[doc = _TAG_ATOMIC!()]
 #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
+/// A boolean type which can be safely shared between threads.
+#[doc = crate::_doc_location!("work/sync/atomic")]
 #[cfg(feature = "dep_portable_atomic")]
 #[cfg_attr(nightly_doc, doc(auto_cfg(hide(feature = "dep_portable_atomic"))))]
 pub use crate::_dep::portable_atomic::AtomicBool;
 //
+#[doc = _TAG_CONCURRENCY!()]
 #[doc = _TAG_ATOMIC!()]
 #[doc = _TAG_ATOMIC_CORE_PORTABLE!()]
+/// A boolean type which can be safely shared between threads.
+#[doc = crate::_doc_location!("work/sync/atomic")]
 #[cfg(not(feature = "dep_portable_atomic"))]
 #[cfg_attr(nightly_doc, doc(auto_cfg(hide(feature = "dep_portable_atomic"))))]
 pub use core::sync::atomic::AtomicBool;